Remy cleaning up the kitchen
